+++ R/tar_target_raw.R
@@ -113,6 +113,9 @@ tar_target_raw <- function(
   )
   tar_assert_format(format)
   tar_assert_repository(repository)
+  if (format == "file_fast" && repository != "local") {
+    format <- "file"
+  }
   tar_assert_flag(iteration, c("vector", "list", "group"))
   tar_assert_flag(
     error,