+++ NEWS.md
@@ -28,6 +28,7 @@ Because of these changes, upgrading to this version of `targets` will unavoidabl
 * For `format = "url"`, only retry on HTTP 500 error codes.
 * Un-deprecate `seconds_interval` and `seconds_timeout` from `tar_resources_url()`, and implement `max_tries` arguments in `tar_resources_aws()` and `tar_resources_gcp()` (#1127).
 * Use `file` and `keep.source` in `parse()` in `callr` utils and target Markdown.
+* Automatically convert `"file_fast"` format to `"file"` format for cloud targets.
 
 # targets 1.2.2